Update patch set 6
Patch Set 6:

> (2 comments)
 > 
 > > (4 comments)
 > >
 > > Still few things not addressed from my last review too.
 > 
 > I've drafted some replies for the comments from the previous patch
 > which they didn't come through because I forgot to click on the
 > Post reply.
 > 
 > I'll have to get back to you about the license and debian package.
 > 
 > Cheers.

Here is a quick copy and paste of my replies to the comments from the Patch 5:

rc-servicecheck/api/resources.py
Line 774:
Can this be put into 1 method, seems duplication for no reason
We needed to separate them into 2 functions for testing.

rc-servicecheck/nectar_monitor_rest/settings.py
Line 29:
are these settings intended to be for running tests or something? What handles setting production settings? Will the debian package handle that?
These are just hard coded for our local dev/testing. We’ll probably have the debian package to handle the settings but we will have to get back to you on that.

Line 99:
What is this for?
This is the email a user will receive from when they unsubscribed from a monitoring list. I've put comment description above it in my patch.

rc-servicecheck/restore_all_configurations.py
Line 12:
Should this be a django management command?
Seems reasonable, should this and the "restore_all_configurations.py" be under the django management command?

Patch-set: 6